Latest additions
Had help from someone on askBSC again with troubleshooting my code. Together we got my quiz running.
The issue was I couldnt get the timer to stop! - I learned that I needed an empty variable in the global space to apply to interval to so that I could call it in the function clearInterval

Afterwards when a user clicked an answer they all came up as incorrect. This was an easy bug I was able to fix on my own.
Now my entire quiz works